+++ backend/budget/handlers/add.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass, field
 
 from budget.models import Budget
 from budget.requests.budget import BudgetRequest
-from budget.responses.budget import BudgetResponse
+from budget.responses.added_budget import AddedBudgetResponse
 from budget.serializer import BudgetSerializer
 from common.handlers.request_manager import RequestManager
 from common.responses.bad_request_response import BadRequestResponse
@@ -17,10 +17,10 @@ from users.models import User
 class AddBudgetRequestManager(RequestManager):
     request: BudgetRequest = field(init=False)
 
-    def safe_process(self) -> BudgetResponse | BadRequestResponse:
+    def safe_process(self) -> AddedBudgetResponse | BadRequestResponse:
         return super().safe_process()  # type: ignore
 
-    def process(self) -> BudgetResponse:
+    def process(self) -> AddedBudgetResponse:
         user: User = self.user_via_token
         budget: Budget = BudgetSerializer().create(validated_data={
             'name': self.request.name,
@@ -28,4 +28,4 @@ class AddBudgetRequestManager(RequestManager):
             'users': [user],
         })
         logging.info(f"New Budget {budget.name}, with {budget.total_budget}. Created by: {user}")
-        return BudgetResponse(budgets=[budget])
+        return AddedBudgetResponse(budgets=[budget])