+++ backend/budget/handlers/get_all.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 
-from budget.responses.budget import BudgetResponse
+from budget.responses.added_budget import AddedBudgetResponse
 from common.handlers.request_manager import RequestManager
 from common.requests.empty import EmptyRequest
 from common.responses.bad_request_response import BadRequestResponse
@@ -14,10 +14,12 @@ from users.models import User
 class GetAllBudgetRequestManager(RequestManager):
     request: EmptyRequest = field(init=False)
 
-    def safe_process(self) -> BudgetResponse | BadRequestResponse:
+    def safe_process(self) -> AddedBudgetResponse | BadRequestResponse:
         return super().safe_process()  # type: ignore
 
-    def process(self) -> BudgetResponse:
+    def process(self) -> AddedBudgetResponse:
         user: User = self.user_via_token
         budgets = user.budgets.all()  # type: ignore
-        return BudgetResponse(budgets)
+        if not budgets:
+            raise NotImplementedError("No budgets found")
+        return AddedBudgetResponse(budgets)