+++ frontend/src/store/budget/creators.ts
@@ -28,7 +28,7 @@ export const creatorsBudgets = {
       dispatch({ type: types.FETCH_REQUESTED });
 
       await BudgetService.add(budgetForm).then((response) => {
-        dispatch({ type: types.FETCH_SUCCESS_APPEND, data: response.data });
+        dispatch({ type: types.FETCH_SUCCESS_NEW, data: response.data });
       });
     };
   },