+++ frontend/src/store/utils/Types.tsx
@@ -3,7 +3,6 @@ export const fetchTypes = (type: string) => {
     FETCH_REQUESTED: `${type}/FETCH_REQUESTED`,
     FETCH_SUCCESS_NEW: `${type}/FETCH_SUCCESS_NEW`,
     FETCH_SUCCESS_TRANSACTION: `${type}/FETCH_SUCCESS_TRANSACTION`,
-    FETCH_SUCCESS_APPEND: `${type}/FETCH_SUCCESS_APPEND`,
     FETCH_ERROR: `${type}/FETCH_ERROR`,
   };
 };