+++ docs/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [[0.7.1]] - 2022-03-26
+
+###### _([diff: 0.7.0-0.7.1])_
+
+### Fixed
+
+- Handling situation when a new budget is added by user who has no budgets added yet.
+- Handling situation when a new transaction is added by user who has no transactions added yet.
+
 ## [[0.7.0]] - 2022-03-26
 
 ###### _([diff: 0.6.3a-0.7.0])_
@@ -378,6 +387,7 @@ Connected Frontend & Backend via webpack_loader.
     -   I know I should implement them right away and have all of that sort of things already done, but I _really_ want to start some dev coding 🐈
     -   ... still got to setup frontend-backend, so I'll prioritize that at this moment.
 
+[diff: 0.7.0-0.7.1]: https://github.com/Luzkan/FamilyBudget/compare/0.7.0...0.7.1
 [diff: 0.6.3a-0.7.0]: https://github.com/Luzkan/FamilyBudget/compare/0.6.3a...0.7.0
 [diff: 0.6.3-0.6.3a]: https://github.com/Luzkan/FamilyBudget/compare/0.6.3...0.6.3a
 [diff: 0.6.2-0.6.3]: https://github.com/Luzkan/FamilyBudget/compare/0.6.2...0.6.3
@@ -405,6 +415,7 @@ Connected Frontend & Backend via webpack_loader.
 [diff: 0.1.1-0.2.0]: https://github.com/Luzkan/FamilyBudget/compare/0.1.1...0.2.0
 [diff: 0.1.0-0.1.1]: https://github.com/Luzkan/FamilyBudget/compare/0.1.0...0.1.1
 [diff: 0.0.0-0.1.0]: https://github.com/Luzkan/FamilyBudget/compare/0.0.0...0.1.0
+[0.7.1]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.7.1
 [0.7.0]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.7.0
 [0.6.3a]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.6.3a
 [0.6.3]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.6.3